Good deal of progress towards a working state + general cleanup
This was 7 distinct commits before, but my laptop overheated and ext4 decided
to corrupt a bunch of stuff under .git, to the point at which git wouldn't even
recognise my local copy as a repository. Thankfully, the actual sources were
unaffected, so all that was lost is history. Pushing this now, in a rather
rough state, to avoid potentially losing work if my laptop decides to fuck up
again.